chore: use xxs mapping for xxxs sizes
