+++ src/Shared/Components/Badge/index.tsx
@@ -1,1 +1,2 @@
 export { default as Badge } from './Badge'
+export type { BadgeProps } from './types'
